refactor: separate usage of @nuxt/kit & disable migrations during dev --remote
