+++ src/runtime/database/server/plugins/migrations.dev.ts
@@ -1,5 +1,5 @@
 import { useRuntimeConfig } from '@nuxt/kit'
-import { applyRemoteMigrations } from '../../../../utils/migrations'
+// import { applyRemoteMigrations } from '../../../../utils/database/migrations'
 import { hubHooks } from '../../../base/server/utils/hooks'
 import { applyMigrations } from '../utils/migrations'
 import { useRuntimeConfig, defineNitroPlugin } from '#imports'
@@ -12,7 +12,7 @@ export default defineNitroPlugin(async () => {
 
   hubHooks.hookOnce('bindings:ready', async () => {
     if (hub.remote && hub.projectKey) { // linked to a NuxtHub project
-      await applyRemoteMigrations()
+      // await applyRemoteMigrations()
     } else { // local dev & self hosted
       await applyMigrations()
     }
